chore: betterer transformer and CLI config APIs; closes #41
## `transformers`

- configurations for a `Transformer` are supplied at instantiation time,
not transform-run time
- created notion of `TransformerConfig`, which is an object having
`id` of `Transformer#id`, and object `config`; instantiate a
`Transformer` with this data using `toTransformer()`
- `validateTransformerChain()` now validates the chain of `Transformer`
objects separately from their instantiation
- this makes `runTransformer()` more simple, as it no longer accepts
any parameters except a source `Observable`.

## `core`

- provide `fromTransformerChain()` for creating `TransformerConfig`
objects, and `transform()` to validate & run the chain

## `cli`

- Use the new core APIs
- Add `commandConfig()` to `commands/common` module, which vastly
simplifies merging of all config options (and a test!)

## Notes

We might want to pass the transformer configuration into the
`outputHeader` and `outputFooter` callbacks of the `table` `Transformer`
because we should be able to dynamically generate these from the config;
right now we have to use function scope to generate the values; e.g.
`diff` command and its `DEFAULT_DIFF_CONFIG`