chore(core): provide consistent public API
I've trimmed down the exports from the `core` package and ensured that
both the default (`Promise`-based) API and the `stream` API export
equivalent functions.

A notable difference is that `builtinRuleDefinitions` is present on the
default API, and it is just an object.  The stream API provides this
as an `Observable` via `fromBuiltinRuleDefinitions`.

These functions are now used from within commands; the `cli` package
no longer depends on `transformers` nor `inspector` packages.

I've tried to make these functions user-friendly, however invoking a
transformer leaves something to be desired.  I chose between two
separate functions--one to initialize the requested transformers,
and another to run them--or a single function which accepts no less than
four (4) parameters:

```js
runTransformer(
  listOfTransformerIds,
  specificTransformerConfigs,
  sourceObservable,
  runTransformerConfig
)
```

The above seemed unwieldy, so I went with two separate functions.  It'd
be possible to allow for the first two parameters above to be a single
configuration object, but that also seems unwieldy.  This is good enough
for now.

Other things:

- Implemented a naïve "raw report object validator", which just checks
  if a given object contains the known top-level properties of a report;
  if a user accidentally passes something that is missing _all_ of these
  objects, we will fail ASAP
- The godlike `fromAny` now detects `Observable` values and passes them
  back out, unmodified
- Added a Sinon stub behavior for returning an RxJS `OperatorFunction`
- When running transforms, the source always becomes multicast.  This
  may be premature optimization, but I had the feeling it was iterating
  over the source `Observable` extra times