feat: external plugin and rule-loading (WIP)
Ref: #32

This adds method `use(pluginId)` to the stream API, which "registers"
a plugin.
A plugin currently is a module (or object) which _may_ export a `rules`
prop, which is an array of `RuleDefinition`s.

`@report-toolkit/inspector/rules/index.js` is such a plugin; likewise
`@report-toolkit/inspector` re-exports `rules`, and the default rules
can be loaded via `use('@report-toolkit/inspector')` by treating it
like a plugin.

The builtin rules themselves can be accessed via the
`fromRegisteredRuleDefinitions()` of the stream API. This will lazy-load
the plugin and the rule definitions (note that the actual `Rule` objects
are not instantiated until it's clear we're going to be using them via
the inspector).

You can theoretically call `use('something')` to load plugin `something`
which would export its own `rules`--the rules would be registered, and
thereafter calling `fromRegisteredRuleDefinitions()` would return these
_in addition to_ the builtin ones.  Currently, this _won't_ give you
the builtins, so I think I need to switch from lazy-loading them to just
loading them outright.