fix: support third-party rules
this re-adds support for third-party rules.  I'm not sure it's the greatest
design, but it's sufficient for now.

if a module id is present in the `plugins` prop of a config, those plugins
will automatically be loaded.  from there, we can reach the rules which it
exports.

this may be going against the grain of the original idea with the config
file, which was to demand a `.js` which requires or imports all of the
external plugins/configurations which it depends upon.  a change to support
that behavior -- instead of just listing module IDs -- would be fairly
trivial to implement.

iirc something must be present in the `rules` prop of the config file
in order for any rules that a plugin exposes to be consumed (having a
non-falsy value here means "enabled")

so the new core APIs are:

1. `use()`, which accepts a module ID
2. `fromRegisteredRuleDefinitions()` now returns an Observable of rule
defs, when combined with configs, can be turned into RuleConfig objects
and run through the inspector
3. util `isPluginRegistered()` and `deregisterPlugin()` methods

need to support these in the `Promise` API, at which point we can close
the issue (see #32)